requirements.txt
slack_bolt
scapy
slack-sdk
allure-pytest==2.13.1
allure-python-commons==2.13.1
apt-xapian-index==0.49
apturl==0.5.2
argcomplete==1.8.1
async-generator==1.10
attrs==22.2.0
ayatana-settings==21.1.28
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cajarename==21.11.24
certifi==2023.5.7
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
deja-dup-caja==0.0.9
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
exceptiongroup==1.1.1
fasteners==0.14.1
folder-color-caja==0.0.86
folder-color-common==0.0.86
future==0.18.2
gpg===1.16.0-unknown
h11==0.14.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
lockfile==0.12.2
louis==3.20.0
Magnus==1.0.3
Mako==1.1.3
MarkupSafe==2.0.1
mate-hud==22.4.4
mate-tweak==22.4.8
modules==1.0.0
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
netifaces==0.11.0
numpy==1.24.3
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
outcome==1.2.0
packaging==23.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pluggy==1.0.0
psutil==5.9.0
ptyprocess==0.7.0
pulsemixer==1.5.1
PyAutoGUI==0.9.53
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
PyGetWindow==0.0.9
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.2.2
pytest-ordering==0.6
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-xapp==2.2.1
python-xlib==0.29
python3-xlib==0.15
pytweening==1.0.4
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
ranger-fm==1.9.3
reportlab==3.6.8
requests==2.25.1
sdkmanager==0.5.1
SecretStorage==3.3.1
selenium==4.9.1
setproctitle==1.2.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
speedtest-cli==2.1.3
ssh-import-id==5.11
systemd-python==234
tesseract==0.1.3
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
wsproto==1.2.0
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zipp==1.0.0
